客户端和服务器实现互相收发数据

+++ QinDevilCommon/SocketList.cs
@@ -47,10 +47,15 @@ namespace QinDevilCommon {
             int i = 0;
             if (data[id] is List<byte>) {
                 List<byte> dataList = data[id] as List<byte>;
-                int v = BitConverter.ToInt32(dataList.GetRange(0, 4).ToArray(), 0);
-                i = BitConverter.ToInt32(dataList.GetRange(4, 4).ToArray(), 0);
-                buffer = dataList.GetRange(8, v - 4).ToArray();
-                dataList.RemoveRange(0, v + 4);
+                if (dataList.Count >= 8) {
+                    byte[] v1 = dataList.GetRange(0, 8).ToArray();
+                    int v = BitConverter.ToInt32(v1, 0);
+                    if (dataList.Count - 4 >= v) {
+                        i = BitConverter.ToInt32(v1, 4);
+                        buffer = dataList.GetRange(8, v - 4).ToArray();
+                        dataList.RemoveRange(0, v + 4);
+                    }
+                }
             }
             return (i, buffer);
         }
